Preparations
* All the code will be rewritten.
* Using dispatch events for reusability.
* **/leave manage** - The staff member will be able to end their LOA’s early and submit extension requests or even subtract time from their own LOA. You will also he able to see your past LOA’s.
* **/loa admin** - Manage other peoples LOA’s and view past LOAs (even if its active)
* **New LOA Embed/Containers** - With discords new v2 components I’ll integrate them to nicely make it pretty. For reference: https://discord.com/channels/1092976553752789054/1092978456196829265/1363220981316976871 (Just an idea).
* **Killing of ban appeals** - Unused, unpractical & a waste of space.

* **Modmail Web Transcripts** - Needed.
* **Reducing the amount of modmail commands.** -  Probably will get rid of modmail snippets and put it in config.
* **Making promotion logs work for promotion and not just edit/delete**
