+++ Cogs/Events/on_loa_ext_request.py
@@ -1,0 +1,1 @@
+# TODO: Logging for extension requests
