+++ Cogs/Events/on_loa_log.py
@@ -1,0 +1,1 @@
+# TODO: Logging for leave modify/create/accept/deny/extend
