+++ Cogs/Events/on_loa_request.py
@@ -1,0 +1,1 @@
+# TODO: Accepting & Denying & Submitting logging and editing.
